condition on radius -- recon loss of 0.17
