+++ burn_vae/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "burn_vae"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[dependencies.pyo3]
+version = "0.19.0"
+default-features = false
+features = ["auto-initialize", "macros"]
+
+[dependencies.once_cell]
+version = "1.19.0"
+default-features = false
+features = ["std"]
+
+[dependencies]
+burn = { workspace = true }
+serde = "1"
+plotly = "0.8.4"
+crossbeam-channel = "0.5.10"
+
