+++ pyproject.toml
@@ -1,0 +1,34 @@
+[tool.poetry]
+name = "burn-vae"
+version = "0.1.0"
+description = "Burn VAE"
+authors = ["Ben Van Sleen"]
+
+[tool.poetry.dependencies]
+python = "3.10.7"
+scikit-learn = "^1.3.2"
+maturin = {extras = ["patchelf"], version = "^1.4.0"}
+plotly = "^5.18.0"
+jupyter = "^1.0.0"
+
+[tool.poetry.group.dev.dependencies]
+jupyter = "^1.0.0"
+
+[build-system]
+requires = ["maturin>=1.4,<2.0"]
+build-backend = "maturin"
+
+[project]
+name = "burn-vae"
+requires-python = ">=3.8"
+classifiers = [
+    "Programming Language :: Rust",
+    "Programming Language :: Python :: Implementation :: CPython",
+    "Programming Language :: Python :: Implementation :: PyPy",
+]
+dynamic = ["version"]
+
+[tool.maturin]
+python-source = "python"
+module-name = "burn_vae._burn_vae"
+features = ["pyo3/extension-module"]
