+++ python/burn_vae/__init__.py
@@ -1,0 +1,1 @@
+from burn_vae._burn_vae import *
